+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/Models/MetricFeedback/MetricChangePointFeedback.cs
@@ -11,7 +11,7 @@ namespace Azure.AI.MetricsAdvisor
     /// Feedback indicating that this is the start of a trend change.
     /// </summary>
     [CodeGenModel("ChangePointFeedback")]
-    [CodeGenSuppress(nameof(MetricChangePointFeedback), typeof(string), typeof(FeedbackDimensionFilter))]
+    [CodeGenSuppress(nameof(MetricChangePointFeedback), typeof(string), typeof(FeedbackFilter))]
     public partial class MetricChangePointFeedback : MetricFeedback
     {
         /// <summary>
@@ -40,12 +40,12 @@ namespace Azure.AI.MetricsAdvisor
 
         /// <summary> Initializes a new <see cref="MetricChangePointFeedback"/> instance. </summary>
         /// <param name="metricId"> The metric unique id. </param>
-        /// <param name="dimensionFilter"> The dimension filter. </param>
+        /// <param name="feedbackFilter"> The dimension filter. </param>
         /// <param name="startTime"> The start timestamp of feedback timerange. </param>
         /// <param name="endTime"> The end timestamp of feedback timerange. When this is equal to <paramref name="startTime"/> it indicates a single timestamp. </param>
         /// <param name="value"> The <see cref="Models.ChangePointFeedbackValue"/> for the feedback. </param>
-        internal MetricChangePointFeedback(string metricId, FeedbackDimensionFilter dimensionFilter, DateTimeOffset startTime, DateTimeOffset endTime, ChangePointFeedbackValue value)
-            : base(metricId, dimensionFilter.DimensionKey)
+        internal MetricChangePointFeedback(string metricId, FeedbackFilter feedbackFilter, DateTimeOffset startTime, DateTimeOffset endTime, ChangePointFeedbackValue value)
+            : base(metricId, feedbackFilter.DimensionKey)
         {
             Argument.AssertNotNull(value, nameof(value));