+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/tests/MetricsAdvisorClient/MetricFeedbackLiveTests.cs
@@ -278,11 +278,14 @@ namespace Azure.AI.MetricsAdvisor.Tests
             var dimensions = new Dictionary<string, string>() { { "city", "Delhi" } };
             var options = new GetAllFeedbackOptions()
             {
-                Filter = new DimensionKey(dimensions),
-                TimeMode = FeedbackQueryTimeMode.FeedbackCreatedOn,
-                StartTime = feedbackSamplingStartTime,
-                EndTime = feedbackSamplingEndTime,
-                FeedbackKind = MetricFeedbackKind.Comment,
+                Filter = new FeedbackFilter()
+                {
+                    DimensionKey = new DimensionKey(dimensions),
+                    TimeMode = FeedbackQueryTimeMode.FeedbackCreatedOn,
+                    StartTime = feedbackSamplingStartTime,
+                    EndTime = feedbackSamplingEndTime,
+                    FeedbackKind = MetricFeedbackKind.Comment,
+                }
             };
 
             var feedbackCount = 0;