+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/tests/Samples/Sample10_FeedbackCrudOperations.cs
@@ -108,9 +108,12 @@ namespace Azure.AI.MetricsAdvisor.Samples
 
             var options = new GetAllFeedbackOptions()
             {
-                StartTime = DateTimeOffset.Parse("2020-01-01T00:00:00Z"),
-                EndTime = DateTimeOffset.Parse("2020-09-09T00:00:00Z"),
-                TimeMode = FeedbackQueryTimeMode.MetricTimestamp,
+                Filter = new FeedbackFilter()
+                {
+                    StartTime = DateTimeOffset.Parse("2020-01-01T00:00:00Z"),
+                    EndTime = DateTimeOffset.Parse("2020-09-09T00:00:00Z"),
+                    TimeMode = FeedbackQueryTimeMode.MetricTimestamp
+                },
                 MaxPageSize = 5
             };